#109: moved oauth2 back to its old place
