+++ http/jsonapi/generator/internal/poi/open-api_test.go
@@ -7,7 +7,7 @@ import (
 	mux "github.com/gorilla/mux"
 	opentracing "github.com/opentracing/opentracing-go"
 	runtime "github.com/pace/bricks/http/jsonapi/runtime"
-	oauth2 "github.com/pace/bricks/http/security/oauth2"
+	oauth2 "github.com/pace/bricks/http/oauth2"
 	errors "github.com/pace/bricks/maintenance/errors"
 	metrics "github.com/pace/bricks/maintenance/metric/jsonapi"
 	"net/http"
